+++ NEWS.md
@@ -2,4 +2,5 @@
 
 * Initial Release
 * Includes support for windows progress bars.
+* Added support for flexible text `R6_box_progress` bars.
 * Added a `NEWS.md` file to track changes to the package.